src/backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.3
python-dotenv==1.0.1
bcrypt==4.1.2
PyJWT==2.8.0
